Refactor alter materialized view statement executor to eliminate code coupling
