+++ fe/fe-core/src/main/java/com/starrocks/alter/AlterJobMgr.java
@@ -36,12 +36,9 @@ package com.starrocks.alter;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
 import com.google.common.collect.Range;
 import com.google.common.collect.Sets;
 import com.starrocks.analysis.DateLiteral;
-import com.starrocks.analysis.IntLiteral;
-import com.starrocks.analysis.StringLiteral;
 import com.starrocks.analysis.TableName;
 import com.starrocks.analysis.TableRef;
 import com.starrocks.authentication.AuthenticationMgr;
@@ -51,7 +48,6 @@ import com.starrocks.catalog.Column;
 import com.starrocks.catalog.DataProperty;
 import com.starrocks.catalog.Database;
 import com.starrocks.catalog.ExpressionRangePartitionInfo;
-import com.starrocks.catalog.ForeignKeyConstraint;
 import com.starrocks.catalog.KeysType;
 import com.starrocks.catalog.MaterializedIndexMeta;
 import com.starrocks.catalog.MaterializedView;
@@ -66,7 +62,6 @@ import com.starrocks.catalog.Table;
 import com.starrocks.catalog.Table.TableType;
 import com.starrocks.catalog.TableProperty;
 import com.starrocks.catalog.Type;
-import com.starrocks.catalog.UniqueConstraint;
 import com.starrocks.catalog.View;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.common.DdlException;
@@ -95,54 +90,40 @@ import com.starrocks.persist.metablock.SRMetaBlockWriter;
 import com.starrocks.privilege.PrivilegeBuiltinConstants;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.ShowResultSet;
-import com.starrocks.scheduler.Constants;
 import com.starrocks.scheduler.Task;
 import com.starrocks.scheduler.TaskBuilder;
-import com.starrocks.scheduler.TaskManager;
-import com.starrocks.scheduler.mv.MaterializedViewMgr;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.server.LocalMetastore;
 import com.starrocks.sql.analyzer.Analyzer;
 import com.starrocks.sql.analyzer.MaterializedViewAnalyzer;
 import com.starrocks.sql.analyzer.SemanticException;
-import com.starrocks.sql.analyzer.SetStmtAnalyzer;
 import com.starrocks.sql.ast.AddPartitionClause;
 import com.starrocks.sql.ast.AlterClause;
-import com.starrocks.sql.ast.AlterMaterializedViewStmt;
+import com.starrocks.sql.ast.AlterMaterializedViewStatusClause;
 import com.starrocks.sql.ast.AlterSystemStmt;
 import com.starrocks.sql.ast.AlterTableCommentClause;
 import com.starrocks.sql.ast.AlterTableStmt;
-import com.starrocks.sql.ast.AsyncRefreshSchemeDesc;
 import com.starrocks.sql.ast.ColumnRenameClause;
 import com.starrocks.sql.ast.CompactionClause;
 import com.starrocks.sql.ast.CreateMaterializedViewStmt;
 import com.starrocks.sql.ast.DropMaterializedViewStmt;
 import com.starrocks.sql.ast.DropPartitionClause;
-import com.starrocks.sql.ast.IntervalLiteral;
 import com.starrocks.sql.ast.ModifyPartitionClause;
 import com.starrocks.sql.ast.ModifyTablePropertiesClause;
 import com.starrocks.sql.ast.PartitionRenameClause;
 import com.starrocks.sql.ast.QueryStatement;
-import com.starrocks.sql.ast.RefreshSchemeDesc;
 import com.starrocks.sql.ast.ReplacePartitionClause;
 import com.starrocks.sql.ast.RollupRenameClause;
-import com.starrocks.sql.ast.SetListItem;
-import com.starrocks.sql.ast.SetStmt;
 import com.starrocks.sql.ast.StatementBase;
-import com.starrocks.sql.ast.SwapTableClause;
-import com.starrocks.sql.ast.SystemVariable;
 import com.starrocks.sql.ast.TableRenameClause;
 import com.starrocks.sql.ast.TruncatePartitionClause;
 import com.starrocks.sql.ast.TruncateTableStmt;
 import com.starrocks.sql.ast.UserIdentity;
-import com.starrocks.sql.common.DmlException;
-import com.starrocks.sql.optimizer.Utils;
 import com.starrocks.sql.optimizer.rule.transformation.materialization.MvUtils;
 import com.starrocks.sql.parser.SqlParser;
 import com.starrocks.thrift.TStorageMedium;
 import com.starrocks.thrift.TTabletMetaType;
 import com.starrocks.thrift.TTabletType;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.threeten.extra.PeriodDuration;
@@ -154,8 +135,6 @@ import java.time.format.DateTimeFormatter;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-
-import static com.starrocks.catalog.TableProperty.INVALID;
 
 public class AlterJobMgr {
     private static final Logger LOG = LogManager.getLogger(AlterJobMgr.class);
@@ -282,103 +261,8 @@ public class AlterJobMgr {
         }
     }
 
-    public void processAlterMaterializedView(AlterMaterializedViewStmt stmt)
-            throws DdlException, MetaNotFoundException {
-        // check db
-        final TableName mvName = stmt.getMvName();
-        final String oldMvName = mvName.getTbl();
-        final String newMvName = stmt.getNewMvName();
-        final RefreshSchemeDesc refreshSchemeDesc = stmt.getRefreshSchemeDesc();
-        final String status = stmt.getStatus();
-        ModifyTablePropertiesClause modifyTablePropertiesClause = stmt.getModifyTablePropertiesClause();
-        String dbName = mvName.getDb();
-        Database db = GlobalStateMgr.getCurrentState().getDb(dbName);
-        if (db == null) {
-            ErrorReport.reportDdlException(ErrorCode.ERR_BAD_DB_ERROR, dbName);
-        }
-        if (!db.writeLockAndCheckExist()) {
-            throw new DdlException("alter materialized failed. database:" + db.getFullName() + " not exist");
-        }
-        MaterializedView materializedView = null;
-        try {
-            final Table table = db.getTable(oldMvName);
-            if (table instanceof MaterializedView) {
-                materializedView = (MaterializedView) table;
-            }
-
-            if (materializedView == null) {
-                throw new MetaNotFoundException("Materialized view " + mvName + " is not found");
-            }
-            // check materialized view state
-            if (materializedView.getState() != OlapTableState.NORMAL) {
-                throw new DdlException("Materialized view [" + materializedView.getName() + "]'s state is not NORMAL. "
-                        + "Do not allow to do ALTER ops");
-            }
-
-            MaterializedViewMgr.getInstance().stopMaintainMV(materializedView);
-
-            // rename materialized view
-            if (newMvName != null) {
-                processRenameMaterializedView(oldMvName, newMvName, db, materializedView);
-            } else if (refreshSchemeDesc != null) {
-                // change refresh scheme
-                processChangeRefreshScheme(refreshSchemeDesc, materializedView, dbName, db);
-            } else if (modifyTablePropertiesClause != null) {
-                try {
-                    processModifyTableProperties(modifyTablePropertiesClause, db, materializedView);
-                } catch (AnalysisException ae) {
-                    throw new DdlException(ae.getMessage());
-                }
-            } else if (status != null) {
-                processChangeMaterializedViewStatus(status, dbName, materializedView, db);
-            } else if (stmt.getSwapTable() != null) {
-                processSwap(db, materializedView, Collections.singletonList(stmt.getSwapTable()));
-            } else {
-                throw new DdlException("Unsupported modification for materialized view");
-            }
-
-            MaterializedViewMgr.getInstance().rebuildMaintainMV(materializedView);
-        } finally {
-            db.writeUnlock();
-        }
-    }
-
-    private void processChangeMaterializedViewStatus(String status, String dbName,
-                                                     MaterializedView materializedView, Database db)
-            throws DdlException, MetaNotFoundException {
-        if (!db.writeLockAndCheckExist()) {
-            throw new DmlException("update meta failed. database:" + db.getFullName() + " not exist");
-        }
-        try {
-            if (AlterMaterializedViewStmt.ACTIVE.equalsIgnoreCase(status)) {
-                if (materializedView.isActive()) {
-                    return;
-                }
-                processChangeMaterializedViewStatus(materializedView, status, false);
-                GlobalStateMgr.getCurrentState().getLocalMetastore()
-                        .refreshMaterializedView(dbName, materializedView.getName(), true, null,
-                                Constants.TaskRunPriority.NORMAL.value(), true, false);
-            } else if (AlterMaterializedViewStmt.INACTIVE.equalsIgnoreCase(status)) {
-                if (!materializedView.isActive()) {
-                    return;
-                }
-                LOG.warn("Setting the materialized view {}({}) to inactive because " +
-                                "user use alter materialized view set status to inactive",
-                        materializedView.getName(), materializedView.getId());
-                processChangeMaterializedViewStatus(materializedView, status, false);
-            } else {
-                throw new DdlException("Unsupported modification materialized view status:" + status);
-            }
-            AlterMaterializedViewStatusLog log = new AlterMaterializedViewStatusLog(materializedView.getDbId(),
-                    materializedView.getId(), status);
-            GlobalStateMgr.getCurrentState().getEditLog().logAlterMvStatus(log);
-        } finally {
-            db.writeUnlock();
-        }
-    }
-
-    private void processChangeMaterializedViewStatus(MaterializedView materializedView, String status, boolean isReplay) {
-        if (AlterMaterializedViewStmt.ACTIVE.equalsIgnoreCase(status)) {
+    public void alterMaterializedViewStatus(MaterializedView materializedView, String status, boolean isReplay) {
+        if (AlterMaterializedViewStatusClause.ACTIVE.equalsIgnoreCase(status)) {
             String viewDefineSql = materializedView.getViewDefineSql();
             ConnectContext context = new ConnectContext();
             context.setQualifiedUser(AuthenticationMgr.ROOT_USER);
@@ -402,7 +286,7 @@ public class AlterJobMgr {
             GlobalStateMgr.getCurrentState().updateBaseTableRelatedMv(materializedView.getDbId(),
                     materializedView, baseTableInfos);
             materializedView.setActive(true);
-        } else if (AlterMaterializedViewStmt.INACTIVE.equalsIgnoreCase(status)) {
+        } else if (AlterMaterializedViewStatusClause.INACTIVE.equalsIgnoreCase(status)) {
             materializedView.setInactiveAndReason("user use alter materialized view set status to inactive");
         }
     }
@@ -414,247 +298,10 @@ public class AlterJobMgr {
         db.writeLock();
         try {
             MaterializedView mv = (MaterializedView) db.getTable(tableId);
-            processChangeMaterializedViewStatus(mv, log.getStatus(), true);
+            alterMaterializedViewStatus(mv, log.getStatus(), true);
         } finally {
             db.writeUnlock();
         }
-    }
-
-    private void processModifyTableProperties(ModifyTablePropertiesClause modifyTablePropertiesClause,
-                                              Database db,
-                                              MaterializedView materializedView) throws AnalysisException {
-        Map<String, String> properties = modifyTablePropertiesClause.getProperties();
-        Map<String, String> propClone = Maps.newHashMap();
-        propClone.putAll(properties);
-        int partitionTTL = INVALID;
-        if (properties.containsKey(PropertyAnalyzer.PROPERTIES_PARTITION_TTL_NUMBER)) {
-            partitionTTL = PropertyAnalyzer.analyzePartitionTimeToLive(properties);
-        }
-        int partitionRefreshNumber = INVALID;
-        if (properties.containsKey(PropertyAnalyzer.PROPERTIES_PARTITION_REFRESH_NUMBER)) {
-            partitionRefreshNumber = PropertyAnalyzer.analyzePartitionRefreshNumber(properties);
-        }
-        String resourceGroup = null;
-        if (properties.containsKey(PropertyAnalyzer.PROPERTIES_RESOURCE_GROUP)) {
-            resourceGroup = PropertyAnalyzer.analyzeResourceGroup(properties);
-            properties.remove(PropertyAnalyzer.PROPERTIES_RESOURCE_GROUP);
-        }
-        int autoRefreshPartitionsLimit = INVALID;
-        if (properties.containsKey(PropertyAnalyzer.PROPERTIES_AUTO_REFRESH_PARTITIONS_LIMIT)) {
-            autoRefreshPartitionsLimit = PropertyAnalyzer.analyzeAutoRefreshPartitionsLimit(properties, materializedView);
-        }
-        List<TableName> excludedTriggerTables = Lists.newArrayList();
-        if (properties.containsKey(PropertyAnalyzer.PROPERTIES_EXCLUDED_TRIGGER_TABLES)) {
-            excludedTriggerTables = PropertyAnalyzer.analyzeExcludedTriggerTables(properties, materializedView);
-        }
-        int maxMVRewriteStaleness = INVALID;
-        if (properties.containsKey(PropertyAnalyzer.PROPERTIES_MV_REWRITE_STALENESS_SECOND)) {
-            maxMVRewriteStaleness = PropertyAnalyzer.analyzeMVRewriteStaleness(properties);
-        }
-        List<UniqueConstraint> uniqueConstraints = Lists.newArrayList();
-        if (properties.containsKey(PropertyAnalyzer.PROPERTIES_UNIQUE_CONSTRAINT)) {
-            uniqueConstraints = PropertyAnalyzer.analyzeUniqueConstraint(properties, db, materializedView);
-            properties.remove(PropertyAnalyzer.PROPERTIES_UNIQUE_CONSTRAINT);
-        }
-        List<ForeignKeyConstraint> foreignKeyConstraints = Lists.newArrayList();
-        if (properties.containsKey(PropertyAnalyzer.PROPERTIES_FOREIGN_KEY_CONSTRAINT)) {
-            foreignKeyConstraints = PropertyAnalyzer.analyzeForeignKeyConstraint(properties, db, materializedView);
-            properties.remove(PropertyAnalyzer.PROPERTIES_FOREIGN_KEY_CONSTRAINT);
-        }
-        TableProperty.QueryRewriteConsistencyMode oldExternalQueryRewriteConsistencyMode =
-                materializedView.getTableProperty().getForceExternalTableQueryRewrite();
-        if (properties.containsKey(PropertyAnalyzer.PROPERTIES_FORCE_EXTERNAL_TABLE_QUERY_REWRITE)) {
-            String propertyValue = properties.get(PropertyAnalyzer.PROPERTIES_FORCE_EXTERNAL_TABLE_QUERY_REWRITE);
-            oldExternalQueryRewriteConsistencyMode = TableProperty.analyzeExternalTableQueryRewrite(propertyValue);
-            properties.remove(PropertyAnalyzer.PROPERTIES_FORCE_EXTERNAL_TABLE_QUERY_REWRITE);
-        }
-        TableProperty.QueryRewriteConsistencyMode oldQueryRewriteConsistencyMode =
-                materializedView.getTableProperty().getQueryRewriteConsistencyMode();
-        if (properties.containsKey(PropertyAnalyzer.PROPERTIES_QUERY_REWRITE_CONSISTENCY)) {
-            String propertyValue = properties.get(PropertyAnalyzer.PROPERTIES_QUERY_REWRITE_CONSISTENCY);
-            oldQueryRewriteConsistencyMode = TableProperty.analyzeQueryRewriteMode(propertyValue);
-            properties.remove(PropertyAnalyzer.PROPERTIES_QUERY_REWRITE_CONSISTENCY);
-        }
-
-        if (!properties.isEmpty()) {
-            if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_COLOCATE_WITH)) {
-                throw new AnalysisException("Modify failed because unsupported properties: " +
-                        "colocate group is not supported for materialized view");
-            }
-            // analyze properties
-            List<SetListItem> setListItems = Lists.newArrayList();
-            for (Map.Entry<String, String> entry : properties.entrySet()) {
-                if (!entry.getKey().startsWith(PropertyAnalyzer.PROPERTIES_MATERIALIZED_VIEW_SESSION_PREFIX)) {
-                    throw new AnalysisException("Modify failed because unknown properties: " + properties +
-                            ", please add `session.` prefix if you want add session variables for mv(" +
-                            "eg, \"session.query_timeout\"=\"30000000\").");
-                }
-                String varKey = entry.getKey().substring(PropertyAnalyzer.PROPERTIES_MATERIALIZED_VIEW_SESSION_PREFIX.length());
-                SystemVariable variable = new SystemVariable(varKey, new StringLiteral(entry.getValue()));
-                setListItems.add(variable);
-            }
-            SetStmtAnalyzer.analyze(new SetStmt(setListItems), null);
-        }
-
-        boolean isChanged = false;
-        Map<String, String> curProp = materializedView.getTableProperty().getProperties();
-        if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_PARTITION_TTL_NUMBER) &&
-                materializedView.getTableProperty().getPartitionTTLNumber() != partitionTTL) {
-            curProp.put(PropertyAnalyzer.PROPERTIES_PARTITION_TTL_NUMBER, String.valueOf(partitionTTL));
-            materializedView.getTableProperty().setPartitionTTLNumber(partitionTTL);
-            isChanged = true;
-        } else if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_PARTITION_REFRESH_NUMBER) &&
-                materializedView.getTableProperty().getPartitionRefreshNumber() != partitionRefreshNumber) {
-            curProp.put(PropertyAnalyzer.PROPERTIES_PARTITION_REFRESH_NUMBER, String.valueOf(partitionRefreshNumber));
-            materializedView.getTableProperty().setPartitionRefreshNumber(partitionRefreshNumber);
-            isChanged = true;
-        } else if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_AUTO_REFRESH_PARTITIONS_LIMIT) &&
-                materializedView.getTableProperty().getAutoRefreshPartitionsLimit() != autoRefreshPartitionsLimit) {
-            curProp.put(PropertyAnalyzer.PROPERTIES_AUTO_REFRESH_PARTITIONS_LIMIT, String.valueOf(autoRefreshPartitionsLimit));
-            materializedView.getTableProperty().setAutoRefreshPartitionsLimit(autoRefreshPartitionsLimit);
-            isChanged = true;
-        } else if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_RESOURCE_GROUP) &&
-                !StringUtils.equals(materializedView.getTableProperty().getResourceGroup(), resourceGroup)) {
-            if (resourceGroup != null && !resourceGroup.isEmpty() &&
-                    GlobalStateMgr.getCurrentState().getResourceGroupMgr().getResourceGroup(resourceGroup) == null) {
-                throw new AnalysisException(PropertyAnalyzer.PROPERTIES_RESOURCE_GROUP
-                        + " " + resourceGroup + " does not exist.");
-            }
-            curProp.put(PropertyAnalyzer.PROPERTIES_RESOURCE_GROUP, resourceGroup);
-            materializedView.getTableProperty().setResourceGroup(resourceGroup);
-            isChanged = true;
-        }
-        if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_EXCLUDED_TRIGGER_TABLES)) {
-            curProp.put(PropertyAnalyzer.PROPERTIES_EXCLUDED_TRIGGER_TABLES,
-                    propClone.get(PropertyAnalyzer.PROPERTIES_EXCLUDED_TRIGGER_TABLES));
-            materializedView.getTableProperty().setExcludedTriggerTables(excludedTriggerTables);
-            isChanged = true;
-        }
-        if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_UNIQUE_CONSTRAINT)) {
-            materializedView.setUniqueConstraints(uniqueConstraints);
-            isChanged = true;
-        }
-        if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_FOREIGN_KEY_CONSTRAINT)) {
-            materializedView.setForeignKeyConstraints(foreignKeyConstraints);
-            isChanged = true;
-        }
-        if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_MV_REWRITE_STALENESS_SECOND)) {
-            curProp.put(PropertyAnalyzer.PROPERTIES_MV_REWRITE_STALENESS_SECOND,
-                    propClone.get(PropertyAnalyzer.PROPERTIES_MV_REWRITE_STALENESS_SECOND));
-            materializedView.setMaxMVRewriteStaleness(maxMVRewriteStaleness);
-            isChanged = true;
-        }
-        if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_FORCE_EXTERNAL_TABLE_QUERY_REWRITE)) {
-            materializedView.getTableProperty().getProperties().
-                    put(PropertyAnalyzer.PROPERTIES_FORCE_EXTERNAL_TABLE_QUERY_REWRITE,
-                            String.valueOf(oldExternalQueryRewriteConsistencyMode));
-            materializedView.getTableProperty().setForceExternalTableQueryRewrite(oldExternalQueryRewriteConsistencyMode);
-            isChanged = true;
-        }
-        if (propClone.containsKey(PropertyAnalyzer.PROPERTIES_QUERY_REWRITE_CONSISTENCY)) {
-            materializedView.getTableProperty().getProperties().
-                    put(PropertyAnalyzer.PROPERTIES_QUERY_REWRITE_CONSISTENCY,
-                            String.valueOf(oldQueryRewriteConsistencyMode));
-            materializedView.getTableProperty().setQueryRewriteConsistencyMode(oldQueryRewriteConsistencyMode);
-            isChanged = true;
-        }
-        DynamicPartitionUtil.registerOrRemovePartitionTTLTable(materializedView.getDbId(), materializedView);
-        if (!properties.isEmpty()) {
-            // set properties if there are no exceptions
-            for (Map.Entry<String, String> entry : properties.entrySet()) {
-                materializedView.getTableProperty().modifyTableProperties(entry.getKey(), entry.getValue());
-            }
-            isChanged = true;
-        }
-
-        if (isChanged) {
-            ModifyTablePropertyOperationLog log = new ModifyTablePropertyOperationLog(materializedView.getDbId(),
-                    materializedView.getId(), propClone);
-            GlobalStateMgr.getCurrentState().getEditLog().logAlterMaterializedViewProperties(log);
-        }
-        LOG.info("alter materialized view properties {}, id: {}", propClone, materializedView.getId());
-    }
-
-    private void processChangeRefreshScheme(RefreshSchemeDesc refreshSchemeDesc, MaterializedView materializedView,
-                                            String dbName, Database db) throws DdlException {
-        MaterializedView.RefreshType newRefreshType = refreshSchemeDesc.getType();
-        MaterializedView.RefreshType oldRefreshType = materializedView.getRefreshScheme().getType();
-
-        TaskManager taskManager = GlobalStateMgr.getCurrentState().getTaskManager();
-        Task currentTask = taskManager.getTask(TaskBuilder.getMvTaskName(materializedView.getId()));
-        Task task;
-        if (currentTask == null) {
-            task = TaskBuilder.buildMvTask(materializedView, dbName);
-            TaskBuilder.updateTaskInfo(task, refreshSchemeDesc, materializedView);
-            taskManager.createTask(task, false);
-        } else {
-            Task changedTask = TaskBuilder.rebuildMvTask(materializedView, dbName, currentTask.getProperties());
-            TaskBuilder.updateTaskInfo(changedTask, refreshSchemeDesc, materializedView);
-            taskManager.alterTask(currentTask, changedTask, false);
-            task = currentTask;
-        }
-
-        // for event triggered type, run task
-        if (task.getType() == Constants.TaskType.EVENT_TRIGGERED) {
-            taskManager.executeTask(task.getName());
-        }
-
-        final MaterializedView.MvRefreshScheme refreshScheme = materializedView.getRefreshScheme();
-        if (!db.writeLockAndCheckExist()) {
-            throw new DmlException("update meta failed. database:" + db.getFullName() + " not exist");
-        }
-        try {
-            // check
-            Table mv = db.getTable(materializedView.getId());
-            if (mv == null) {
-                throw new DmlException(
-                        "update meta failed. materialized view:" + materializedView.getName() + " not exist");
-            }
-            refreshScheme.setType(newRefreshType);
-            if (refreshSchemeDesc instanceof AsyncRefreshSchemeDesc) {
-                AsyncRefreshSchemeDesc asyncRefreshSchemeDesc = (AsyncRefreshSchemeDesc) refreshSchemeDesc;
-                IntervalLiteral intervalLiteral = asyncRefreshSchemeDesc.getIntervalLiteral();
-                if (intervalLiteral != null) {
-                    final IntLiteral step = (IntLiteral) intervalLiteral.getValue();
-                    final MaterializedView.AsyncRefreshContext asyncRefreshContext = refreshScheme.getAsyncRefreshContext();
-                    if (asyncRefreshSchemeDesc.isDefineStartTime()) {
-                        asyncRefreshContext.setStartTime(Utils.getLongFromDateTime(asyncRefreshSchemeDesc.getStartTime()));
-                    }
-                    asyncRefreshContext.setStep(step.getLongValue());
-                    asyncRefreshContext.setTimeUnit(intervalLiteral.getUnitIdentifier().getDescription());
-                } else {
-                    if (materializedView.getBaseTableInfos().stream().anyMatch(tableInfo ->
-                            !tableInfo.getTable().isNativeTableOrMaterializedView()
-                    )) {
-                        throw new DdlException("Materialized view which type is ASYNC need to specify refresh interval for " +
-                                "external table");
-                    }
-                    refreshScheme.setAsyncRefreshContext(new MaterializedView.AsyncRefreshContext());
-                }
-            }
-
-            final ChangeMaterializedViewRefreshSchemeLog log = new ChangeMaterializedViewRefreshSchemeLog(materializedView);
-            GlobalStateMgr.getCurrentState().getEditLog().logMvChangeRefreshScheme(log);
-        } finally {
-            db.writeUnlock();
-        }
-        LOG.info("change materialized view refresh type {} to {}, id: {}", oldRefreshType,
-                newRefreshType, materializedView.getId());
-    }
-
-    private void processRenameMaterializedView(String oldMvName, String newMvName, Database db,
-                                               MaterializedView materializedView) throws DdlException {
-        if (db.getTable(newMvName) != null) {
-            throw new DdlException("Materialized view [" + newMvName + "] is already used");
-        }
-        materializedView.setName(newMvName);
-        db.dropTable(oldMvName);
-        db.registerTableUnlocked(materializedView);
-        final RenameMaterializedViewLog renameMaterializedViewLog =
-                new RenameMaterializedViewLog(materializedView.getId(), db.getId(), newMvName);
-        updateTaskDefinition(materializedView);
-        GlobalStateMgr.getCurrentState().getEditLog().logMvRename(renameMaterializedViewLog);
-        LOG.info("rename materialized view[{}] to {}, id: {}", oldMvName, newMvName, materializedView.getId());
     }
 
     public void replayRenameMaterializedView(RenameMaterializedViewLog log) {
@@ -843,7 +490,7 @@ public class AlterJobMgr {
             } else if (currentAlterOps.hasRenameOp()) {
                 processRename(db, olapTable, alterClauses);
             } else if (currentAlterOps.hasSwapOp()) {
-                processSwap(db, olapTable, alterClauses);
+                new AlterJobExecutor().process(stmt, ConnectContext.get());
             } else if (currentAlterOps.hasAlterCommentOp()) {
                 processAlterComment(db, olapTable, alterClauses);
             } else if (currentAlterOps.contains(AlterOpType.MODIFY_TABLE_PROPERTY_SYNC)) {
@@ -948,63 +595,18 @@ public class AlterJobMgr {
         }
     }
 
-    // entry of processing swap table
-    private void processSwap(Database db, OlapTable origTable, List<AlterClause> alterClauses) throws DdlException {
-        if (!(alterClauses.get(0) instanceof SwapTableClause)) {
-            throw new DdlException("swap operation only support table");
-        }
-
-        // must hold db write lock
-        Preconditions.checkState(db.isWriteLockHeldByCurrentThread());
-
-        SwapTableClause clause = (SwapTableClause) alterClauses.get(0);
-        String origTblName = origTable.getName();
-        String newTblName = clause.getTblName();
-        Table newTbl = db.getTable(newTblName);
-        if (newTbl == null || !(newTbl.isOlapOrCloudNativeTable() || newTbl.isMaterializedView())) {
-            throw new DdlException("Table " + newTblName + " does not exist or is not OLAP/LAKE table");
-        }
-        OlapTable olapNewTbl = (OlapTable) newTbl;
-
-        // First, we need to check whether the table to be operated on can be renamed
-        olapNewTbl.checkAndSetName(origTblName, true);
-        origTable.checkAndSetName(newTblName, true);
-
-        if (origTable.isMaterializedView() || newTbl.isMaterializedView()) {
-            if (!(origTable.isMaterializedView() && newTbl.isMaterializedView())) {
-                throw new DdlException("Materialized view can only SWAP WITH materialized view");
-            }
-        }
-
-        // inactive the related MVs
-        LocalMetastore.inactiveRelatedMaterializedView(db, origTable,
-                String.format("based table %s swapped", origTblName));
-        LocalMetastore.inactiveRelatedMaterializedView(db, olapNewTbl,
-                String.format("based table %s swapped", newTblName));
-
-        swapTableInternal(db, origTable, olapNewTbl);
-
-        // write edit log
-        SwapTableOperationLog log = new SwapTableOperationLog(db.getId(), origTable.getId(), olapNewTbl.getId());
-        GlobalStateMgr.getCurrentState().getEditLog().logSwapTable(log);
-
-        LOG.info("finish swap table {}-{} with table {}-{}", origTable.getId(), origTblName, newTbl.getId(),
-                newTblName);
-    }
-
     public void replaySwapTable(SwapTableOperationLog log) {
+        try {
+            swapTableInternal(log);
+        } catch (DdlException e) {
+            LOG.warn("should not happen", e);
+        }
         long dbId = log.getDbId();
         long origTblId = log.getOrigTblId();
         long newTblId = log.getNewTblId();
         Database db = GlobalStateMgr.getCurrentState().getDb(dbId);
         OlapTable origTable = (OlapTable) db.getTable(origTblId);
         OlapTable newTbl = (OlapTable) db.getTable(newTblId);
-
-        try {
-            swapTableInternal(db, origTable, newTbl);
-        } catch (DdlException e) {
-            LOG.warn("should not happen", e);
-        }
         LOG.debug("finish replay swap table {}-{} with table {}-{}", origTblId, origTable.getName(), newTblId,
                 newTbl.getName());
     }
@@ -1014,8 +616,14 @@ public class AlterJobMgr {
      * For example, SWAP TABLE A WITH TABLE B.
      * must pre check A can be renamed to B and B can be renamed to A
      */
-    private void swapTableInternal(Database db, OlapTable origTable, OlapTable newTbl)
-            throws DdlException {
+    public void swapTableInternal(SwapTableOperationLog log) throws DdlException {
+        long dbId = log.getDbId();
+        long origTblId = log.getOrigTblId();
+        long newTblId = log.getNewTblId();
+        Database db = GlobalStateMgr.getCurrentState().getDb(dbId);
+        OlapTable origTable = (OlapTable) db.getTable(origTblId);
+        OlapTable newTbl = (OlapTable) db.getTable(newTblId);
+
         String origTblName = origTable.getName();
         String newTblName = newTbl.getName();
