+++ fe/fe-core/src/main/java/com/starrocks/alter/AlterOpType.java
@@ -61,6 +61,8 @@ public enum AlterOpType {
 
     //Alter View
     ALTER_VIEW,
+    REFRESH_SCHEMA,
+    ALTER_MV_STATUS,
 
     INVALID_OP; // INVALID_OP must be the last one
