+++ fe/fe-core/src/main/java/com/starrocks/common/util/PropertyAnalyzer.java
@@ -64,6 +64,7 @@ import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.server.RunMode;
 import com.starrocks.sql.analyzer.AnalyzerUtils;
+import com.starrocks.sql.analyzer.SemanticException;
 import com.starrocks.sql.ast.Property;
 import com.starrocks.thrift.TCompressionType;
 import com.starrocks.thrift.TStorageMedium;
@@ -283,16 +284,16 @@ public class PropertyAnalyzer {
         return shortKeyColumnCount;
     }
 
-    public static int analyzePartitionTimeToLive(Map<String, String> properties) throws AnalysisException {
+    public static int analyzePartitionTimeToLive(Map<String, String> properties) {
         int partitionTimeToLive = INVALID;
         if (properties != null && properties.containsKey(PROPERTIES_PARTITION_TTL_NUMBER)) {
             try {
                 partitionTimeToLive = Integer.parseInt(properties.get(PROPERTIES_PARTITION_TTL_NUMBER));
             } catch (NumberFormatException e) {
-                throw new AnalysisException("Partition TTL Number: " + e.getMessage());
+                throw new SemanticException("Partition TTL Number: " + e.getMessage());
             }
             if (partitionTimeToLive <= 0 && partitionTimeToLive != INVALID) {
-                throw new AnalysisException("Illegal Partition TTL Number: " + partitionTimeToLive);
+                throw new SemanticException("Illegal Partition TTL Number: " + partitionTimeToLive);
             }
             properties.remove(PROPERTIES_PARTITION_TTL_NUMBER);
         }
@@ -318,10 +319,9 @@ public class PropertyAnalyzer {
         return partitionLiveNumber;
     }
 
-    public static int analyzeAutoRefreshPartitionsLimit(Map<String, String> properties, MaterializedView mv)
-            throws AnalysisException {
+    public static int analyzeAutoRefreshPartitionsLimit(Map<String, String> properties, MaterializedView mv) {
         if (mv.getRefreshScheme().getType() == MaterializedView.RefreshType.MANUAL) {
-            throw new AnalysisException(
+            throw new SemanticException(
                     "The auto_refresh_partitions_limit property does not support manual refresh mode.");
         }
         int autoRefreshPartitionsLimit = -1;
@@ -329,36 +329,35 @@ public class PropertyAnalyzer {
             try {
                 autoRefreshPartitionsLimit = Integer.parseInt(properties.get(PROPERTIES_AUTO_REFRESH_PARTITIONS_LIMIT));
             } catch (NumberFormatException e) {
-                throw new AnalysisException("Auto Refresh Partitions Limit: " + e.getMessage());
+                throw new SemanticException("Auto Refresh Partitions Limit: " + e.getMessage());
             }
             if (autoRefreshPartitionsLimit <= 0 && autoRefreshPartitionsLimit != INVALID) {
-                throw new AnalysisException("Illegal Auto Refresh Partitions Limit: " + autoRefreshPartitionsLimit);
+                throw new SemanticException("Illegal Auto Refresh Partitions Limit: " + autoRefreshPartitionsLimit);
             }
             properties.remove(PROPERTIES_AUTO_REFRESH_PARTITIONS_LIMIT);
         }
         return autoRefreshPartitionsLimit;
     }
 
-    public static int analyzePartitionRefreshNumber(Map<String, String> properties) throws AnalysisException {
+    public static int analyzePartitionRefreshNumber(Map<String, String> properties) {
         int partitionRefreshNumber = -1;
         if (properties != null && properties.containsKey(PROPERTIES_PARTITION_REFRESH_NUMBER)) {
             try {
                 partitionRefreshNumber = Integer.parseInt(properties.get(PROPERTIES_PARTITION_REFRESH_NUMBER));
             } catch (NumberFormatException e) {
-                throw new AnalysisException("Partition Refresh Number: " + e.getMessage());
+                throw new SemanticException("Partition Refresh Number: " + e.getMessage());
             }
             if (partitionRefreshNumber <= 0 && partitionRefreshNumber != INVALID) {
-                throw new AnalysisException("Illegal Partition Refresh Number: " + partitionRefreshNumber);
+                throw new SemanticException("Illegal Partition Refresh Number: " + partitionRefreshNumber);
             }
             properties.remove(PROPERTIES_PARTITION_REFRESH_NUMBER);
         }
         return partitionRefreshNumber;
     }
 
-    public static List<TableName> analyzeExcludedTriggerTables(Map<String, String> properties, MaterializedView mv)
-            throws AnalysisException {
+    public static List<TableName> analyzeExcludedTriggerTables(Map<String, String> properties, MaterializedView mv) {
         if (mv.getRefreshScheme().getType() != MaterializedView.RefreshType.ASYNC) {
-            throw new AnalysisException("The excluded_trigger_tables property only applies to asynchronous refreshes.");
+            throw new SemanticException("The excluded_trigger_tables property only applies to asynchronous refreshes.");
         }
         List<TableName> tables = Lists.newArrayList();
         if (properties != null && properties.containsKey(PROPERTIES_EXCLUDED_TRIGGER_TABLES)) {
@@ -369,7 +368,7 @@ public class PropertyAnalyzer {
                 if (mv.containsBaseTable(tableName)) {
                     tables.add(tableName);
                 } else {
-                    throw new AnalysisException(tableName.toSql() +
+                    throw new SemanticException(tableName.toSql() +
                             " is not base table of materialized view " + mv.getName());
                 }
             }
@@ -378,17 +377,16 @@ public class PropertyAnalyzer {
         return tables;
     }
 
-    public static int analyzeMVRewriteStaleness(Map<String, String> properties)
-            throws AnalysisException {
+    public static int analyzeMVRewriteStaleness(Map<String, String> properties) {
         int maxMVRewriteStaleness = INVALID;
         if (properties != null && properties.containsKey(PROPERTIES_MV_REWRITE_STALENESS_SECOND)) {
             try {
                 maxMVRewriteStaleness = Integer.parseInt(properties.get(PROPERTIES_MV_REWRITE_STALENESS_SECOND));
             } catch (NumberFormatException e) {
-                throw new AnalysisException("Invalid maxMVRewriteStaleness Number: " + e.getMessage());
+                throw new SemanticException("Invalid maxMVRewriteStaleness Number: " + e.getMessage());
             }
             if (maxMVRewriteStaleness != INVALID && maxMVRewriteStaleness < 0) {
-                throw new AnalysisException("Illegal maxMVRewriteStaleness: " + maxMVRewriteStaleness);
+                throw new SemanticException("Illegal maxMVRewriteStaleness: " + maxMVRewriteStaleness);
             }
             properties.remove(PROPERTIES_MV_REWRITE_STALENESS_SECOND);
         }
@@ -714,8 +712,7 @@ public class PropertyAnalyzer {
         return val;
     }
 
-    public static List<UniqueConstraint> analyzeUniqueConstraint(
-            Map<String, String> properties, Database db, OlapTable table) throws AnalysisException {
+    public static List<UniqueConstraint> analyzeUniqueConstraint(Map<String, String> properties, Database db, OlapTable table) {
         List<UniqueConstraint> uniqueConstraints = Lists.newArrayList();
         List<UniqueConstraint> analyzedUniqueConstraints = Lists.newArrayList();
 
@@ -726,7 +723,7 @@ public class PropertyAnalyzer {
             }
             uniqueConstraints = UniqueConstraint.parse(uniqueConstraintStr);
             if (uniqueConstraints == null || uniqueConstraints.isEmpty()) {
-                throw new AnalysisException(String.format("invalid unique constraint:%s", uniqueConstraintStr));
+                throw new SemanticException(String.format("invalid unique constraint:%s", uniqueConstraintStr));
             }
 
             for (UniqueConstraint uniqueConstraint : uniqueConstraints) {
@@ -736,19 +733,19 @@ public class PropertyAnalyzer {
                     String dbName = uniqueConstraint.getDbName() != null ? uniqueConstraint.getDbName()
                             : db.getFullName();
                     if (uniqueConstraint.getTableName() == null) {
-                        throw new AnalysisException("must set table name for unique constraint in materialized view");
+                        throw new SemanticException("must set table name for unique constraint in materialized view");
                     }
                     String tableName = uniqueConstraint.getTableName();
                     Table uniqueConstraintTable = GlobalStateMgr.getCurrentState().getMetadataMgr().getTable(catalogName,
                             dbName, tableName);
                     if (uniqueConstraintTable == null) {
-                        throw new AnalysisException(
+                        throw new SemanticException(
                                 String.format("table: %s.%s.%s does not exist", catalogName, dbName, tableName));
                     }
                     boolean columnExist = uniqueConstraint.getUniqueColumns().stream()
                             .allMatch(uniqueConstraintTable::containColumn);
                     if (!columnExist) {
-                        throw new AnalysisException(
+                        throw new SemanticException(
                                 String.format("some columns of:%s do not exist in table:%s.%s.%s",
                                         uniqueConstraint.getUniqueColumns(), catalogName, dbName, tableName));
                     }
@@ -757,7 +754,7 @@ public class PropertyAnalyzer {
                 } else {
                     boolean columnExist = uniqueConstraint.getUniqueColumns().stream().allMatch(table::containColumn);
                     if (!columnExist) {
-                        throw new AnalysisException(
+                        throw new SemanticException(
                                 String.format("some columns of:%s do not exist in table:%s",
                                         uniqueConstraint.getUniqueColumns(), table.getName()));
                     }
@@ -771,8 +768,7 @@ public class PropertyAnalyzer {
 
     private static Pair<BaseTableInfo, Table> analyzeForeignKeyConstraintTablePath(String tablePath,
                                                                                    String foreignKeyConstraintDesc,
-                                                                                   Database db)
-            throws AnalysisException {
+                                                                                   Database db) {
         String[] parts = tablePath.split("\\.");
         String catalogName = InternalCatalog.DEFAULT_INTERNAL_CATALOG_NAME;
         String dbName = db.getFullName();
@@ -787,22 +783,22 @@ public class PropertyAnalyzer {
         } else if (parts.length == 1) {
             tableName = parts[0];
         } else {
-            throw new AnalysisException(String.format("invalid foreign key constraint:%s," +
+            throw new SemanticException(String.format("invalid foreign key constraint:%s," +
                     "table path is invalid", foreignKeyConstraintDesc));
         }
 
         if (!GlobalStateMgr.getCurrentState().getCatalogMgr().catalogExists(catalogName)) {
-            throw new AnalysisException(String.format("catalog: %s do not exist", catalogName));
+            throw new SemanticException(String.format("catalog: %s do not exist", catalogName));
         }
         Database parentDb = GlobalStateMgr.getCurrentState().getMetadataMgr().getDb(catalogName, dbName);
         if (parentDb == null) {
-            throw new AnalysisException(
+            throw new SemanticException(
                     String.format("catalog: %s, database: %s do not exist", catalogName, dbName));
         }
         Table table = GlobalStateMgr.getCurrentState().getMetadataMgr()
                 .getTable(catalogName, dbName, tableName);
         if (table == null) {
-            throw new AnalysisException(String.format("catalog:%s, database: %s, table:%s do not exist",
+            throw new SemanticException(String.format("catalog:%s, database: %s, table:%s do not exist",
                     catalogName, dbName, tableName));
         }
 
@@ -817,8 +813,7 @@ public class PropertyAnalyzer {
     }
 
     private static void analyzeForeignKeyUniqueConstraint(Table parentTable, List<String> parentColumns,
-                                                          Table analyzedTable)
-            throws AnalysisException {
+                                                          Table analyzedTable) {
         KeysType parentTableKeyType = KeysType.DUP_KEYS;
         if (parentTable.isNativeTableOrMaterializedView()) {
             OlapTable parentOlapTable = (OlapTable) parentTable;
@@ -834,12 +829,12 @@ public class PropertyAnalyzer {
         }
 
         if (parentTableKeyType == KeysType.AGG_KEYS) {
-            throw new AnalysisException(
+            throw new SemanticException(
                     String.format("do not support reference agg table:%s", parentTable.getName()));
         } else if (parentTableKeyType == KeysType.DUP_KEYS) {
             // for DUP_KEYS type olap table or external table
             if (!parentTable.hasUniqueConstraints() && mvUniqueConstraints.isEmpty()) {
-                throw new AnalysisException(
+                throw new SemanticException(
                         String.format("dup table:%s has no unique constraint", parentTable.getName()));
             } else {
                 List<UniqueConstraint> uniqueConstraints = parentTable.getUniqueConstraints();
@@ -856,7 +851,7 @@ public class PropertyAnalyzer {
                     }
                 }
                 if (!matched) {
-                    throw new AnalysisException(
+                    throw new SemanticException(
                             String.format("columns:%s are not dup table:%s's unique constraint", parentColumns,
                                     parentTable.getName()));
                 }
@@ -865,14 +860,14 @@ public class PropertyAnalyzer {
             // for PRIMARY_KEYS and UNIQUE_KEYS type table
             // parent columns should be keys
             if (!((OlapTable) parentTable).isKeySet(Sets.newHashSet(parentColumns))) {
-                throw new AnalysisException(String.format("columns:%s are not key columns of table:%s",
+                throw new SemanticException(String.format("columns:%s are not key columns of table:%s",
                         parentColumns, parentTable.getName()));
             }
         }
     }
 
     public static List<ForeignKeyConstraint> analyzeForeignKeyConstraint(
-            Map<String, String> properties, Database db, Table analyzedTable) throws AnalysisException {
+            Map<String, String> properties, Database db, Table analyzedTable) {
         List<ForeignKeyConstraint> foreignKeyConstraints = Lists.newArrayList();
         if (properties != null && properties.containsKey(PROPERTIES_FOREIGN_KEY_CONSTRAINT)) {
             String foreignKeyConstraintsDesc = properties.get(PROPERTIES_FOREIGN_KEY_CONSTRAINT);
@@ -888,7 +883,7 @@ public class PropertyAnalyzer {
                 }
                 Matcher foreignKeyMatcher = ForeignKeyConstraint.FOREIGN_KEY_PATTERN.matcher(trimed);
                 if (!foreignKeyMatcher.find() || foreignKeyMatcher.groupCount() != 9) {
-                    throw new AnalysisException(
+                    throw new SemanticException(
                             String.format("invalid foreign key constraint:%s", foreignKeyConstraintDesc));
                 }
                 String sourceTablePath = foreignKeyMatcher.group(1);
@@ -902,7 +897,7 @@ public class PropertyAnalyzer {
                 List<String> parentColumns = Arrays.stream(targetColumns.split(",")).
                         map(String::trim).map(String::toLowerCase).collect(Collectors.toList());
                 if (childColumns.size() != parentColumns.size()) {
-                    throw new AnalysisException(String.format("invalid foreign key constraint:%s," +
+                    throw new SemanticException(String.format("invalid foreign key constraint:%s," +
                             " columns' size does not match", foreignKeyConstraintDesc));
                 }
                 // analyze table exist for foreign key constraint
@@ -911,7 +906,7 @@ public class PropertyAnalyzer {
                 BaseTableInfo parentTableInfo = parentTablePair.first;
                 Table parentTable = parentTablePair.second;
                 if (!parentColumns.stream().allMatch(parentTable::containColumn)) {
-                    throw new AnalysisException(String.format("some columns of:%s do not exist in parent table:%s",
+                    throw new SemanticException(String.format("some columns of:%s do not exist in parent table:%s",
                             parentColumns, parentTable.getName()));
                 }
 
@@ -922,15 +917,15 @@ public class PropertyAnalyzer {
                             db);
                     childTable = childTablePair.second;
                     if (!childColumns.stream().allMatch(childTable::containColumn)) {
-                        throw new AnalysisException(String.format("some columns of:%s do not exist in table:%s",
+                        throw new SemanticException(String.format("some columns of:%s do not exist in table:%s",
                                 childColumns, childTable.getName()));
                     }
                 } else {
                     if (!analyzedTable.isNativeTable()) {
-                        throw new AnalysisException("do not support add foreign key on external table");
+                        throw new SemanticException("do not support add foreign key on external table");
                     }
                     if (!childColumns.stream().allMatch(analyzedTable::containColumn)) {
-                        throw new AnalysisException(String.format("some columns of:%s do not exist in table:%s",
+                        throw new SemanticException(String.format("some columns of:%s do not exist in table:%s",
                                 childColumns, analyzedTable.getName()));
                     }
                 }
@@ -943,7 +938,7 @@ public class PropertyAnalyzer {
                     Column childColumn = childTable.getColumn(pair.first);
                     Column parentColumn = parentTable.getColumn(pair.second);
                     if (!childColumn.getType().equals(parentColumn.getType())) {
-                        throw new AnalysisException(String.format(
+                        throw new SemanticException(String.format(
                                 "column:%s type does mot match referenced column:%s type", pair.first, pair.second));
                     }
                 }
@@ -954,7 +949,7 @@ public class PropertyAnalyzer {
                 foreignKeyConstraints.add(foreignKeyConstraint);
             }
             if (foreignKeyConstraints.isEmpty()) {
-                throw new AnalysisException(
+                throw new SemanticException(
                         String.format("invalid foreign key constrain:%s", foreignKeyConstraintsDesc));
             }
             properties.remove(PROPERTIES_FOREIGN_KEY_CONSTRAINT);
