+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/AlterMaterializedViewStmt.java
@@ -15,11 +15,8 @@
 
 package com.starrocks.sql.ast;
 
-import com.google.common.collect.Sets;
 import com.starrocks.analysis.TableName;
 import com.starrocks.sql.parser.NodePosition;
-
-import java.util.Set;
 
 /**
  * 1.Support for modifying the way of refresh and the cycle of asynchronous refresh;
@@ -27,59 +24,21 @@ import java.util.Set;
  * 3.SYNC is not supported and ASYNC is not allow changed to SYNC
  */
 public class AlterMaterializedViewStmt extends DdlStmt {
-
     private final TableName mvName;
-    private final String newMvName;
-    private final RefreshSchemeDesc refreshSchemeDesc;
-    private final ModifyTablePropertiesClause modifyTablePropertiesClause;
-    private final String status;
-    private final SwapTableClause swapTable;
+    private final AlterTableClause alterTableClause;
 
-    public static final String ACTIVE = "active";
-    public static final String INACTIVE = "inactive";
-    public static final Set<String> SUPPORTED_MV_STATUS = Sets.newTreeSet(String.CASE_INSENSITIVE_ORDER);
-
-    static {
-        SUPPORTED_MV_STATUS.add(ACTIVE);
-        SUPPORTED_MV_STATUS.add(INACTIVE);
-    }
-
-    public AlterMaterializedViewStmt(TableName mvName, String newMvName,
-                                     RefreshSchemeDesc refreshSchemeDesc,
-                                     ModifyTablePropertiesClause modifyTablePropertiesClause,
-                                     String status, SwapTableClause swapTable,
-                                     NodePosition pos) {
+    public AlterMaterializedViewStmt(TableName mvName, AlterTableClause alterTableClause, NodePosition pos) {
         super(pos);
         this.mvName = mvName;
-        this.newMvName = newMvName;
-        this.refreshSchemeDesc = refreshSchemeDesc;
-        this.modifyTablePropertiesClause = modifyTablePropertiesClause;
-        this.status = status;
-        this.swapTable = swapTable;
+        this.alterTableClause = alterTableClause;
     }
 
     public TableName getMvName() {
         return mvName;
     }
 
-    public String getNewMvName() {
-        return newMvName;
-    }
-
-    public RefreshSchemeDesc getRefreshSchemeDesc() {
-        return refreshSchemeDesc;
-    }
-
-    public ModifyTablePropertiesClause getModifyTablePropertiesClause() {
-        return modifyTablePropertiesClause;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public SwapTableClause getSwapTable() {
-        return swapTable;
+    public AlterTableClause getAlterTableClause() {
+        return alterTableClause;
     }
 
     @Override
